src/main/java/com/github/joschi/jadconfig/jodatime/converters/DaysConverter.java
package com.github.joschi.jadconfig.jodatime.converters;

import com.github.joschi.jadconfig.Converter;
import com.github.joschi.jadconfig.ParameterException;
import org.joda.time.Days;

/**
 * Converter for type {@link org.joda.time.Days}
 *
 * @author jschalanda
 */
public class DaysConverter implements Converter<Days> {

    /**
     * Returns a {@link org.joda.time.Days} instance representing the specified {@link String} {@literal value}.
     *
     * @param value The configuration parameter's {@link String} value
     * @return A {@link org.joda.time.Days} instance representing the configuration parameter's value
     */
    @Override
    public Days convertFrom(String value) {
        try {
            final String trimmed = value.trim();
            if (trimmed.startsWith("P")) {
                return Days.parseDays(trimmed);
            } else {
                final int days = Integer.parseInt(trimmed);
                return Days.days(days);
            }
        } catch (Exception ex) {
            throw new ParameterException("Couldn't convert value \"" + value + "\" to Days.", ex);
        }
    }

    /**
     * Returns a {@link String} instance representing the configuration parameter's {@literal value}.
     *
     * @param value The configuration parameter's {@link org.joda.time.Days} value
     * @return A {@link String} instance representing the configuration parameter's typed value
     */
    @Override
    public String convertTo(Days value) {
        if (value == null) {
            throw new ParameterException("Couldn't convert \"null\" to String.");
        }

        return String.valueOf(value.getDays());
    }
}
